[HumanBenchmark][Backend][Seed] add the API fetch in the seed file
