+++ back-end/prisma/seed.js
@@ -1,5 +1,5 @@
 const { PrismaClient } = require("@prisma/client");
-
+const axios = require("axios");
 const prisma = new PrismaClient();
 
 
@@ -71,7 +71,42 @@ const usersData = [
 ];
 
 
+async function fetchAndSeedExercises() {
+  const options = {
+    method: "GET",
+    url: "https://exercisedb.p.rapidapi.com/exercises",
+    params: { limit: "0", offset: "0" },
+    headers: {
+      "x-rapidapi-key": process.env.VITE_RAPIDAPI_KEY,
+      "x-rapidapi-host": "exercisedb.p.rapidapi.com",
+    },
+  };
+
+  try {
+    const response = await axios.request(options);
+    const transformedData = response.data.map((exercise) => ({
+      name: exercise.name,
+      bodyParts: [exercise.bodyPart],
+      targetMuscle: exercise.target,
+      overview: exercise.description,
+      exerciseTips: exercise.instructions.join(" "),
+      createdAt: new Date(),
+    }));
+
+    for (const exercise of transformedData) {
+      await prisma.exercise.create({ data: exercise });
+    }
+
+    console.log("Exercises seeded successfully!");
+  } catch (error) {
+    console.error("Error fetching and seeding exercises:", error);
+  }
+}
+
+
 async function main() {
+  await fetchAndSeedExercises();
+
   // seed users
   for (const userData of usersData) {
     const { workouts, muscleStats, bodyPartStats, ...user } = userData;
