+++ front-end/src/ExerciseContext.jsx
@@ -8,36 +8,9 @@ export const ExerciseProvider = ({ children }) => {
 
   useEffect(() => {
     const fetchExercises = async () => {
-      const options = {
-        method: "GET",
-        url: "https://exercisedb.p.rapidapi.com/exercises",
-        params: { limit: "0", offset: "0" },
-        headers: {
-          "x-rapidapi-key": import.meta.env.VITE_RAPIDAPI_KEY,
-          "x-rapidapi-host": "exercisedb.p.rapidapi.com",
-        },
-      };
-
       try {
-        const response = await axios.request(options);
-        const transformedData = response.data.map((exercise) => ({
-          //id is provided from the API, but we need to generate our own unique id
-          name: exercise.name,
-          bodyParts: [exercise.bodyPart],
-          targetMuscle: exercise.target,
-          overview: exercise.description,
-          exerciseTips: exercise.instructions.join(" "),
-          createdAt: new Date(),
-        }));
-
-        //console.log("Transformed Data:", transformedData);
-
-      for (const exercise of transformedData) {
-        await axios.post("http://localhost:3000/api/exercises", [exercise]);
-      }
-
-      const exercisesResponse = await axios.get("http://localhost:3000/api/exercises");
-      setExercises(exercisesResponse.data);
+        const exercisesResponse = await axios.get("http://localhost:3000/api/exercises");
+        setExercises(exercisesResponse.data);
       } catch (error) {
         console.error("Error fetching exercises:", error);
       }
